Added local test for create token requset
