--- src/test/java/co/omise/resources/TokenResourceTest.java
@@ -1,42 +1,0 @@
-package co.omise.resources;
-
-import co.omise.Endpoint;
-import co.omise.models.Card;
-import co.omise.models.OmiseException;
-import co.omise.models.Token;
-import org.joda.time.Period;
-import org.joda.time.YearMonth;
-import org.junit.Test;
-
-import java.io.IOException;
-
-public class TokenResourceTest extends ResourceTest {
-    private static final String TOKEN_ID = "tokn_test_4yq8lbecl0q6dsjzxr5";
-
-    @Test
-    public void testCreate() throws IOException, OmiseException {
-        Token token = resource().create(new Token.Create().card(new Card.Create()
-                .name("JOHN DOE")
-                .number("4242424242424242")
-                .expiration(YearMonth.now().withPeriodAdded(Period.years(1), 1))
-                .securityCode("123")
-                .city("Bangkok")
-                .postalCode("10240"))
-        );
-
-        assertRequested("POST", "/tokens", 200);
-        assertVaultRequest();
-
-        assertEquals(TOKEN_ID, token.getId());
-        assertFalse(token.isLiveMode());
-        assertEquals("card_test_4yq6tuucl9h4erukfl0", token.getCard().getId());
-    }
-
-    private void assertVaultRequest() {
-        assertEquals(Endpoint.VAULT.host(), lastRequest().url().host());
-    }
-
-    private TokenResource resource() {
-        return new TokenResource(testClient());
-    }
-}
